# backend/service.core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entity.Table:
    properties:
      num_seats:
        type: integer
      restaurant_id:
        type: string
      status:
        type: string
      table_num:
        type: integer
    type: object
...
  errors.ErrorResponse:
    properties:
      message:
        example: error message
        type: string
    type: object
...
  tables.CreateTableRequest:
    properties:
      num_seats:
        type: integer
      table_num:
        type: integer
    type: object
  tables.ItemResponse:
    properties:
      item:
        $ref: '#/definitions/entity.Table'
        type: object
    type: object
  tables.UpdateTableRequest:
    properties:
      num_seats:
        type: integer
      status:
        type: string
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: Restaurants and customers and other great things!
  license: {}
  title: COMP3900 - JAMAR
  version: "0.1"
paths:
...
  /restaurants/{res_id}/tables:
    get:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/tables.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Gets all tables associated to a restaurant
      tags:
      - Tables
    post:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      - description: The table's details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tables.CreateTableRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/tables.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Creates a new table for a restaurant
      tags:
      - Tables
  /restaurants/{res_id}/tables/{tbl_num}:
    get:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      - description: The id of the table
        in: path
        name: tbl_num
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/tables.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get a table's details
      tags:
      - Tables
    put:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      - description: The id of the table
        in: path
        name: tbl_num
        required: true
        type: string
      - description: The table's updated details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/tables.UpdateTableRequest'
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/tables.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Updates a table for a restaurant
      tags:
      - Tables
  /restaurants/{res_id}/tables/{tbl_num}/free:
    post:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      - description: The id of the table
        in: path
        name: tbl_num
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/tables.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Frees a table
      tags:
      - Tables
  /restaurants/{res_id}/tables/{tbl_num}/occupy:
    post:
      parameters:
      - description: The id of the restaurant
        in: path
        name: res_id
        required: true
        type: string
      - description: The id of the table
        in: path
        name: tbl_num
        required: true
        type: string
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/tables.ItemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Occupies a table
      tags:
      - Tables
...
swagger: "2.0"